refactor: refactoring create user logic
