+++ src/core/repositories/users-repository.ts
@@ -1,6 +1,6 @@
 import type { User } from "../entities/user";
 
-export interface CreateUserData {
+export interface RegisterUserData {
   name: string;
   email: string;
   password: string;
@@ -10,14 +10,14 @@ export interface AuthUserData {
   email: string;
   password: string;
 }
-export interface CreateUserDataRepo {
+export interface RegisterUserDataRepo {
   name: string;
   email: string;
   password_hash: string;
 }
 
 export interface UsersRepository {
-  create(data: CreateUserDataRepo): Promise<User>;
+  create(data: RegisterUserDataRepo): Promise<User>;
   findByEmail(email: string): Promise<User | null>;
   findById(id: string): Promise<User | null>;
 }
