+++ src/core/use-cases/factories/make-register-use-case.ts
@@ -1,0 +1,9 @@
+import { MethodsUsersRepository } from "@/repositories/db-users-repository";
+import { RegisterUseCase } from "../users/register";
+
+export function makeRegisterUseCase() {
+  const usersRepository = new MethodsUsersRepository();
+  const registerUseCase = new RegisterUseCase(usersRepository);
+
+  return registerUseCase;
+}
